+++ wss-20211221/ChangeLog.md
@@ -1,3 +1,15 @@
+2025-02-07 Version: 1.1.0
+- Support API ModifyInstanceProperties.
+- Update API DescribePackageDeductions: add param EndTime.
+- Update API DescribePackageDeductions: add param StartTime.
+- Update API DescribePackageDeductions: update param InstanceIds.
+- Update API DescribePackageDeductions: update param PackageIds.
+- Update API DescribePackageDeductions: update param PageNum.
+- Update API DescribePackageDeductions: update param PageSize.
+- Update API DescribePackageDeductions: update param ResourceType.
+- Update API DescribePackageDeductions: update response param.
+
+
 2024-10-03 Version: 1.0.0
 - Generated python 2021-12-21 for wss.